feat(util): add size parameter to extractThumbnail method
The code changes in `sl1_thumbnail.dart` add a new optional `size` parameter to the `extractThumbnail` method in the `ThumbnailUtil` class. This allows the caller to specify the desired size of the thumbnail when retrieving the file from the API. The method now accepts the `size` parameter and passes it to the `getFileThumbnail` method in the `ApiService` class.